+++ src/models/TsDeclaration.ts
