+++ src/models/TsResolveFile.ts
@@ -1,6 +1,6 @@
 import {TsFile} from './TsFile';
 import {TsImport} from './TsImport';
-import {TsExport} from './TsExport';
+import {TsExport} from './TsDeclaration';
 
 export class TsResolveFile extends TsFile {
     public imports: TsImport[] = [];
